lua: extract heap implementation to lua_heap.cpp file
