+++ firmware/controllers/lua/lua.cpp
@@ -6,184 +6,20 @@
 #if EFI_LUA
 
 #include "lua.hpp"
+#include "lua_heap.h"
 #include "lua_hooks.h"
 #include "can_filter.h"
 
 #define TAG "LUA "
 
 static bool withErrorLoading = false;
+static int luaTickPeriodUs;
 
-#if EFI_PROD_CODE || EFI_SIMULATOR
-
-#ifndef LUA_USER_HEAP
-// At least one heap_header_t should fit
-// see [tag:multi-step-lua-alloc] below
-// this is a bit over-complicated at the moment, one argument would be that this supports multi-region RAM use-case
-#define LUA_USER_HEAP 16
-#endif // LUA_USER_HEAP
-
-//#ifdef PERSISTENT_LOCATION_TODO
-//#define LUA_HEAP_RAM_SECTION CCM_OPTIONAL
-//#endif
-
-CH_HEAP_AREA(luaUserHeap, LUA_USER_HEAP)
-#ifdef EFI_HAS_EXT_SDRAM
-SDRAM_OPTIONAL
-#endif
-#ifdef LUA_HEAP_RAM_SECTION
-LUA_HEAP_RAM_SECTION
-#endif
-;
-
+#if EFI_CAN_SUPPORT
 static int recentRxCount = 0;
 static int totalRxCount = 0;
 static int rxTime;
-
-
-class Heap {
-public:
-	memory_heap_t m_heap;
-
-	size_t m_size = 0;
-	uint8_t* m_buffer = nullptr;
-
-	void* alloc(size_t n) {
-		if (m_buffer && m_size) {
-			return chHeapAlloc(&m_heap, n);
-		}
-
-		return nullptr;
-	}
-
-	void free(void* obj) {
-		chHeapFree(obj);
-	}
-
-public:
-	template<size_t TSize>
-	Heap(uint8_t (&buffer)[TSize])
-	{
-		reinit(buffer, TSize);
-	}
-
-	void reinit(uint8_t *buffer, size_t size) {
-		criticalAssertVoid(used() == 0, "Too late to reinit Lua heap");
-
-		m_size = size;
-		m_buffer = buffer;
-
-		reset();
-	}
-
-	void* realloc(void* ptr, size_t osize, size_t nsize) {
-		if (nsize == 0) {
-			// requested size is zero, free if necessary and return nullptr
-			if (ptr) {
-				free(ptr);
-			}
-
-			return nullptr;
-		}
-
-		void *new_mem = nullptr;
-
-		// try this heap first
-		new_mem = alloc(nsize);
-
-    // [tag:multi-step-lua-alloc]
-		// then try ChibiOS default heap
-		if (new_mem == nullptr) {
-			new_mem = chHeapAlloc(NULL, nsize);
-		}
-
-		if (!ptr) {
-			// No old pointer passed in, simply return allocated block
-			return new_mem;
-		}
-
-		// An old pointer was passed in, copy the old data in, then free
-		if (new_mem != nullptr) {
-			memcpy(new_mem, ptr, chHeapGetSize(ptr) > nsize ? nsize : chHeapGetSize(ptr));
-			free(ptr);
-		}
-
-		return new_mem;
-	}
-
-	size_t size() const {
-		return m_size;
-	}
-
-	size_t used() {
-		if (size() == 0) {
-			return 0;
-		}
-
-		size_t heapFree = 0;
-		size_t lagestFree = 0;
-		chHeapStatus(&m_heap, &heapFree, &lagestFree);
-		// hack to return zero when heap is totaly free
-		// this is for leak detector
-		// if all free memory is in one chunk this means heap is totaly free
-		if (heapFree == lagestFree) {
-			return 0;
-		}
-		return m_size - heapFree;
-	}
-
-	// Use only in case of emergency - obliterates all heap objects and starts over
-	void reset() {
-		if (m_buffer && m_size) {
-			chHeapObjectInit(&m_heap, m_buffer, m_size);
-		}
-	}
-};
-
-static Heap userHeap(luaUserHeap);
-
-static void printLuaMemoryInfo() {
-	auto heapSize = userHeap.size();
-
-	if (heapSize) {
-		auto memoryUsed = userHeap.used();
-		float pct = 100.0f * memoryUsed / heapSize;
-		efiPrintf("Dedicated Lua memory heap usage: %d / %d bytes = %.1f%%", memoryUsed, heapSize, pct);
-	} else {
-		efiPrintf("No dedicated Lua heap, using ChibiOS default heap");
-	}
-
-	size_t chHeapFree = 0;
-	chHeapStatus(NULL, &chHeapFree, NULL);
-	efiPrintf("Common ChibiOS heap: %d bytes free", chHeapFree);
-	efiPrintf("ChibiOS memcore: %d bytes free", chCoreGetStatusX());
-}
-
-static void* myAlloc(void* /*ud*/, void* optr, size_t osize, size_t nsize) {
-	void *nptr = userHeap.realloc(optr, osize, nsize);
-
-	if (engineConfiguration->debugMode == DBG_LUA) {
-		engine->outputChannels.debugIntField1 = userHeap.used();
-	}
-
-	return nptr;
-}
-#else // not EFI_PROD_CODE
-// Non-MCU code can use plain realloc function instead of custom implementation
-static void* myAlloc(void* /*ud*/, void* ptr, size_t /*osize*/, size_t nsize) {
-	if (!nsize) {
-		free(ptr);
-		return nullptr;
-	}
-
-	if (!ptr) {
-		return malloc(nsize);
-	}
-
-	return realloc(ptr, nsize);
-}
-#endif // EFI_PROD_CODE
-
-static int luaTickPeriodUs;
+#endif // EFI_CAN_SUPPORT
 
 static int lua_setTickRate(lua_State* l) {
 	float freq = luaL_checknumber(l, 1);
@@ -261,7 +97,7 @@ static bool loadScript(LuaHandle& ls, const char* scriptStr) {
 	efiPrintf(TAG "script loaded successfully!");
 
 #if EFI_PROD_CODE
-	printLuaMemoryInfo();
+	luaHeapPrintInfo();
 #endif // EFI_PROD_CODE
 
 	return true;
@@ -401,9 +237,9 @@ static bool runOneLua(lua_Alloc alloc, const char* script) {
 
 void LuaThread::ThreadTask() {
 	while (!chThdShouldTerminateX()) {
-		bool wasOk = runOneLua(myAlloc, config->luaScript);
+		bool wasOk = runOneLua(luaHeapAlloc, config->luaScript);
 
-		auto usedAfterRun = userHeap.used();
+		auto usedAfterRun = luaHeapUsed();
 		if (usedAfterRun != 0) {
 		  if (!withErrorLoading) {
 			  efiPrintf(TAG "MEMORY LEAK DETECTED: %d bytes used after teardown", usedAfterRun);
@@ -411,7 +247,7 @@ void LuaThread::ThreadTask() {
 
 			// Lua blew up in some terrible way that left memory allocated, reset the heap
 			// so that subsequent runs don't overflow the heap
-			userHeap.reset();
+			luaHeapReset();
 		}
 
 		// Reset any lua adjustments the script made
@@ -481,7 +317,7 @@ void startLua() {
 	  efiPrintf("luaCycle %luus including luaRxTime %dus", NT2US(engine->outputChannels.luaLastCycleDuration),
 	    NT2US(rxTime));
 
-     printLuaMemoryInfo();
+     luaHeapPrintInfo();
   });
 #endif
 }
@@ -494,7 +330,7 @@ void startLua() { }
 #include <string>
 
 static LuaHandle runScript(const char* script) {
-	auto ls = setupLuaState(myAlloc);
+	auto ls = setupLuaState(luaHeapAlloc);
 
 	if (!ls) {
 		throw std::logic_error("Call to setupLuaState failed, returned null");
@@ -560,7 +396,7 @@ int testLuaReturnsInteger(const char* script) {
 }
 
 void testLuaExecString(const char* script) {
-	auto ls = setupLuaState(myAlloc);
+	auto ls = setupLuaState(luaHepAlloc);
 
 	if (!ls) {
 		throw std::logic_error("Call to setupLuaState failed, returned null");
