+++ firmware/controllers/lua/lua.mk
@@ -2,6 +2,7 @@ LUA_DIR=$(CONTROLLERS_DIR)/lua
 LUA_EXT=$(PROJECT_DIR)/ext/lua
 
 ALLCPPSRC += $(LUA_DIR)/lua.cpp \
+			 $(LUA_DIR)/lua_heap.cpp \
 			 $(LUA_DIR)/lua_hooks.cpp \
 			 $(LUA_DIR)/can_filter.cpp \
 			 $(LUA_DIR)/lua_hooks_util.cpp \
