+++ firmware/controllers/lua/lua_heap.h
@@ -1,0 +1,9 @@
+// file lua_heap.h
+
+#pragma once
+
+void* luaHeapAlloc(void* /*ud*/, void* optr, size_t osize, size_t nsize);
+size_t luaHeapUsed();
+void luaHeapReset();
+
+void luaHeapPrintInfo();
